+++ src/main/kotlin/io/kjson/JSONArray.kt
@@ -2,7 +2,7 @@
  * @(#) JSONArray.kt
  *
  * kjson-core  JSON Kotlin core functionality
- * Copyright (c) 2021, 2022 Peter Wall
+ * Copyright (c) 2021, 2022, 2023 Peter Wall
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -45,7 +45,7 @@ import net.pwall.util.output
 class JSONArray internal constructor (array: Array<out JSONValue?>, override val size: Int) : JSONStructure<Int>,
         List<JSONValue?> {
 
-    private val immutableList = ImmutableList<JSONValue?>(array, size)
+    internal val immutableList = ImmutableList<JSONValue?>(array, size)
 
     override fun appendTo(a: Appendable) {
         a.append('[')
@@ -60,6 +60,8 @@ class JSONArray internal constructor (array: Array<out JSONValue?>, override val
         }
         a.append(']')
     }
+
+    override fun toJSON(): String = if (isEmpty()) "[]" else buildString { appendTo(this) }
 
     override fun output(out: IntConsumer) {
         out.accept('['.code)
@@ -115,6 +117,14 @@ class JSONArray internal constructor (array: Array<out JSONValue?>, override val
         if (fromIndex == 0)
             return JSONArray(oldArray, toIndex)
         return JSONArray(oldArray.copyOfRange(fromIndex, toIndex), toIndex - fromIndex)
+    }
+
+    fun forEachItem(func: (JSONValue?) -> Unit) {
+        repeat(size) { func(immutableList[it]) }
+    }
+
+    fun forEachItemIndexed(func: (Int, JSONValue?) -> Unit) {
+        repeat(size) { func(it, immutableList[it]) }
     }
 
     @Suppress("SuspiciousEqualsCombination")